Refactoring the SharedInfo class to "APrioriInfo" to avoid confusion
Renaming SharedInfo to APrioriInfo as it is easily confused with the term
"Shared Secret" of the ECDH which is completely different. The APrioriInfo
is information that is communicated prior to the key exchange event taking
place and is used in place of a CA (the next best thing given the circumstances)
as part of the ECG key exchange protocol.