Attempted to improve security by using both pieces of shared info
This is a temporary fix to help improve the security of the ECG key
exchange by using both piece of the a priori shared information to
sign & verify the public keys (S1 + S2) instead of just S1 or S2
depending on if you are the initiator or recipient.

This is still a security flaw, as the weakest part of the cryptosystem
is the shared information that the users specify. If an attacker breaks
the shared information then there are multiple attack vectors on the
cryptosystem. To put it bluntly, if an attacker breaks the shared
information used and somehow has access to the public keys the entire
cryptosystem is broken.

A potential solution may be to integrate gesture input for entering
the shared information S1 & S2.